# Implement window operators in c++