--- a/temporian/core/operators/window/base.py
+++ b/temporian/core/operators/window/base.py
@@ -15,10 +15,11 @@
 """Base calendar operator."""
 
 from abc import ABC, abstractmethod
-from typing import Optional
+from typing import Optional, List
 
 
 from temporian.core.data.duration import Duration
+from temporian.core.data import dtype
 from temporian.core.data.event import Event
 from temporian.core.data.feature import Feature
 from temporian.core.data.sampling import Sampling
@@ -44,22 +45,24 @@ class BaseWindowOperator(Operator, ABC):
 
         if sampling is not None:
             self.add_input("sampling", sampling)
+            self._has_sampling = True
             effective_sampling = sampling.sampling()
         else:
             effective_sampling = event.sampling()
+            self._has_sampling = False
 
         self.add_input("event", event)
 
-        # TODO: Remve auto prefix
         output_features = [  # pylint: disable=g-complex-comprehension
             Feature(
-                name=f"{self.prefix}_{f.name()}",
+                name=f.name(),
                 dtype=self.get_feature_dtype(f),
                 sampling=effective_sampling,
                 creator=self,
             )
             for f in event.features()
         ]
+        self._output_dtypes = [feature.dtype() for feature in output_features]
 
         # output
         self.add_output(
@@ -73,8 +76,14 @@ class BaseWindowOperator(Operator, ABC):
 
         self.check()
 
-    def window_length(self):
+    def window_length(self) -> Duration:
         return self._window_length
+
+    def has_sampling(self) -> bool:
+        return self._has_sampling
+
+    def output_dtypes(self) -> List[dtype.DType]:
+        return self._output_dtypes
 
     @classmethod
     def build_op_definition(cls) -> pb.OperatorDef:
@@ -110,8 +119,3 @@ class BaseWindowOperator(Operator, ABC):
         Returns:
             str: The dtype of the output feature.
         """
-
-    @property
-    @abstractmethod
-    def prefix(self) -> str:
-        """Get the prefix to use for the output features."""