--- a/temporian/implementation/numpy/operators/window/base.py
+++ b/temporian/implementation/numpy/operators/window/base.py
@@ -12,11 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import ABC, abstractmethod
-from typing import Dict, Optional, List
+from abc import abstractmethod
+from typing import Dict, Optional, List, Any
 
 import numpy as np
-from temporian.core.data.duration import Duration
 from temporian.core.operators.window.base import BaseWindowOperator
 from temporian.implementation.numpy.data.event import NumpyEvent
 from temporian.implementation.numpy.data.event import NumpyFeature
@@ -54,6 +53,10 @@ class BaseWindowNumpyImplementation(OperatorImplementation):
 
         return {"event": dst_event}
 
+    @abstractmethod
+    def _implementation(self) -> Any:
+        pass
+
     def _compute(
         self,
         src_timestamps: np.ndarray,
@@ -61,78 +64,14 @@ class BaseWindowNumpyImplementation(OperatorImplementation):
         sampling_timestamps: np.ndarray,
         dst_features: List[NumpyFeature],
     ):
-        mask = self._build_accumulator_mask(
-            src_timestamps,
-            sampling_timestamps,
-            self.operator.window_length(),
-        )
-
-        # For each feature
+        implementation = self._implementation()
         for src_ts in src_features:
-            dst_feature_name = src_ts.name
-            dst_ts_data = self._apply_accumulator_mask(src_ts.data, mask)
-            dst_features.append(NumpyFeature(dst_feature_name, dst_ts_data))
-
-    def _build_accumulator_mask(
-        self,
-        data_timestamps: np.array,
-        sampling_timestamps: np.array,
-        win: Duration,
-    ) -> np.array:
-        """Creates a 2d boolean matrix containing the summing instructions.
-
-        The returned matrix "m[i,j]" is defined as:
-            m[i,j] is true iif. input value "j" is averaged in the output value "i".
-        """
-
-        # This implementation is simple but expensive. It will create multiple
-        # O(n^2) arrays, where n is the number of time samples.
-
-        right_side = (
-            sampling_timestamps[:, np.newaxis] >= data_timestamps[np.newaxis, :]
-        )
-
-        # TODO: Make left side inclusivity/exclusivity a parameter.
-        left_side = sampling_timestamps[:, np.newaxis] <= (
-            data_timestamps[np.newaxis, :] + win
-        )
-
-        return right_side & left_side
-
-    def _apply_accumulator_mask(
-        self, src: np.array, mask: np.array
-    ) -> np.array:
-        """Sums elements according to an accumulator mask."""
-
-        # Broadcast of feature values to requested timestamps.
-        cross_product = src * mask
-
-        # Replace masked values with NaN
-        cross_product[mask == False] = (
-            np.nan
-        )  # pylint: disable=singleton-comparison
-
-        return self._apply_operation(cross_product)
-
-    @abstractmethod
-    def _apply_operation(self, values: np.array) -> np.array:
-        """
-        Applies a window operator to each of the values in each row of the
-        input array.
-
-        The input array should have a shape (n, m), where 'n' is the length of
-        the feature and 'm' is the size of the window. Each row represents a
-        window of data points, with 'nan' values used for padding when the
-        window size is smaller than the number of data points in the time
-        series. The function should compute the operation for each row (window).
-
-        Args:
-            values: A 2D NumPy array with shape (n, m) where each row represents
-                a  window of data points. 'n' is the length of the feature, and
-                'm' is the size of the window. The array can contain 'nan'
-                values as padding.
-
-        Returns:
-            np.array: A 1D NumPy array with shape (n,) containing the operation
-                    for each row (window) in the input array.
-        """
+            args = {
+                "event_timestamps": src_timestamps,
+                "event_values": src_ts.data,
+                "window_length": self.operator.window_length(),
+            }
+            if self.operator.has_sampling():
+                args["sampling_timestamps"] = sampling_timestamps
+            dst_feature = implementation(**args)
+            dst_features.append(NumpyFeature(src_ts.name, dst_feature))