--- a/temporian/implementation/numpy/operators/window/moving_count.py
+++ b/temporian/implementation/numpy/operators/window/moving_count.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import numpy as np
 
 from temporian.implementation.numpy import implementation_lib
 from temporian.core.operators.window.moving_count import (
@@ -22,37 +21,14 @@ from temporian.implementation.numpy.operators.window.base import (
     BaseWindowNumpyImplementation,
 )
 
+from temporian.implementation.numpy_cc.operators import window as window_cc
+
 
 class MovingCountNumpyImplementation(BaseWindowNumpyImplementation):
     """Numpy implementation of the moving count operator."""
 
-    def __init__(self, operator: MovingCountOperator) -> None:
-        super().__init__(operator)
-
-    def _apply_operation(self, values: np.array) -> np.array:
-        """
-        Calculates the non-nan count of the values in each row of the input
-        array.
-
-        The input array should have a shape (n, m), where 'n' is the length of
-        the feature and 'm' is the size of the window. Each row represents a
-        window of data points, with 'nan' values used for padding when the
-        window size is  smaller than the number of data points in the time
-        series. The function  computes the count for each row (window) while
-        ignoring the 'nan' values.
-
-        Args:
-            values: A 2D NumPy array with shape (n, m) where each row represents
-                a  window of data points. 'n' is the length of the feature, and
-                'm' is the size of the window. The array can contain 'nan'
-                values as padding.
-
-        Returns:
-            np.array: A 1D NumPy array with shape (n,) containing the non-nan
-                    count for each row (window) in the input array.
-
-        """
-        return np.count_nonzero(~np.isnan(values), axis=1).astype(np.int64)
+    def _implementation(self):
+        return window_cc.moving_count
 
 
 implementation_lib.register_operator_implementation(