… .vscode/settings.json
… README.md
… temporian/implementation/numpy/operators/test/BUILD
… temporian/implementation/numpy/operators/test/moving_standard_deviation_test.py
… temporian/implementation/numpy/operators/test/moving_sum_test.py
… temporian/implementation/numpy/operators/test/simple_moving_average_test.py
… temporian/implementation/numpy_cc/operators/window.cc
… temporian/test/api_test.py