Bonus:
- Using the benchmark script, we see a x2000 speed-up.
- There is not allocated work memory anymore (only the output).

Benchmark:

Before
```
Running lag benchmark with N=300000...
Build schedule
Run 1 operators
    1 / 1: SIMPLE_MOVING_AVERAGE [7.26928 s]
Execution in 7.26963 s
```

After
```
Running lag benchmark with N=300000...
Build schedule
Run 1 operators
    1 / 1: SIMPLE_MOVING_AVERAGE [0.00371 s]
Execution in 0.00405 s
```